requirements.txt
# FCL-PRM: Federated Continual Process Reward Model
# Python 3.11+

# Core ML
torch>=2.2.0
transformers>=4.40.0
datasets>=2.19.0
accelerate>=0.30.0

# Data & IO
numpy>=1.26.0
pyyaml>=6.0.1
tqdm>=4.66.0

# Evaluation
scikit-learn>=1.5.0
scipy>=1.13.0

# Experiment tracking (optional)
wandb>=0.17.0

# Privacy (P4)
opacus>=1.5.0

# Development
pytest>=8.2.0
black>=24.4.0
isort>=5.13.0
mypy>=1.10.0

# Jupyter for analysis
jupyter>=1.0.0
matplotlib>=3.9.0
seaborn>=0.13.0
